RUBY-1733 Modify test_parent_notified_of_child_start to see if it will pass on Travis
For some inexplicable reason the AbstractSegmentTest#test_parent_notified_of_child_start
test case fails on TravisCI, but passes locally and using the Travis docker image. I've
updated the test to check the same functionality, but to avoid Mocha expects, which I
suspect contributes to this issue. The changes I had to make are pretty terrible, so
I likely won't keep the test if it passes this time around. I'm just interested to see
if it this fixes things.