RUBY-1733 Remove test_parent_notified_of_child_start
This test was failing on Travis when using a mocha expectation,
but not locally or in the Travis VM. I rewrote in the last commit
to avoid using a mocha expectation, but that required using
instance_variable_get. The test passed on Travis, ensuring that
the behavior is correct, but in it's current state this test
is not very useful. It's better off to remove it. This functionality
is tested indirectly in a number of situations.