RUBY-1856 Rename request_uri => request.uri
There are currently agents in the wild that record an attribute under
both names. The request.uri form is more consistent with the other
agent attributes. Since some agents will have to change the name
either way, we should go with the one that will be more consistent
going forward.