RUBY-1876 Introduce PrioritySampledBuffer class
This commit introduces the PrioritySampledBuffer the class so that we can
keep events by a randomly assigned priority number. Similar to our SampledBuffer
we want the option to accept a block so that we do not need to create events
that won't be saved. To that end, the append method expects to receive either
a priority and block, or a single event element with a priority key. In the
former case, the block will only be evaluated if an event with that priority
will be admitted to the buffer.