RUBY-1927 Add custom_attributes configuration option
This option will correspond to the custom_parameters option that
was added for LASP. We call our configuration option custom_attributes
instead of custom_parameters to remain consistent with our APIs. For
example, we have NewRelic::Agent.add_custom_attributes. This API used
to be named add_custom_parameters, but it was deprecated and later
removed.