RUBY-1989 Initial test case structure and metrics verification
The verify_metrics method call is commented out due to a handful of
failing tests. Many of the issues have to do with us not validating
certain fields in the payload, and hence, will require some code changes
to fix, the driver is otherwise working fine. It is still incomplete, however.