RUBY-2031 Perf tests comparing Transaction#start with Tracer#start_transaction
With the Transaction#start test we contiue the tradition of first
looking up the transaction state, then passing it to all the methods
that need access to it. In the Tracer#start_transaction test, we rely
on methods exposed off of the Tracer that independently look up information
from the state when needed. The perf tests show that both methods of starting
a transaction are equivalent from a performance perspective. Below is the
output of the perf tests:

TransactionTracingPerfTests#test_start_with_transaction_start: 8313.517 ips (120.29 µs / iteration)
    cpu_time_user:        4.98 (0.00 / iter)
  cpu_time_system:        0.08 (0.00 / iter)
          gc_runs:          48 (0.00 / iter)
     live_objects:         -37 (-0.00 / iter)
      allocations: 2.54625e+06 (60.58 / iter)
       iterations:       42029

TransactionTracingPerfTests#test_start_with_tracer_start: 8307.016 ips (120.38 µs / iteration)
    cpu_time_user:           5 (0.00 / iter)
  cpu_time_system:        0.08 (0.00 / iter)
          gc_runs:          48 (0.00 / iter)
     live_objects:         173 (0.00 / iter)
      allocations: 2.55802e+06 (60.58 / iter)
       iterations:       42228