RUBY-2031 Remove end_time parameter from Transaction.stop
This parameter is unused in the codebase, and furthermore, it incurs an additional
object allocation for each call to transaction stop. The only time the end_time is used
is during the last call to stop. Here are the results from running the basic middleware
perf test:

mwear:ruby_agent mwear$ be ruby test/performance/script/runner -n test_basic_middleware_stack -C
1 tests, 0 failures, 7.711108 s total
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
| name                                       | before    | after     | delta | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 476.47 µs | 464.30 µs | -2.6% |           332 |          321 |        -3.3% |
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+