RUBY-2031 Make Transaction#stop behavior equivalent to Transaction.stop
Ultimately, the API we are striving for is that transactions are started
via Tracer#start_transaction, and they are finished via Transaction#stop
(the instance method). This commit removes paramters from Transaction#stop
and has it delegate to Transaction.stop in the event of multiple calls
to stop. Ultimately, we want to remove the frame_stack and multiple calls
to stop, we are taking an incremental approach, and those changes will
come in future commits.