RUBY-2023 Remove end_time parameter from Transaction#record_apdex
There were a handful of tests that relied on being able to call
record_apdex with an end_time for the purposes of recording apdex
metrics for satisfying, tolerating, frustrating values. Those test
cases have been expanded so that there is a transaction per desired
apdex (which is how this would play out in the real world).