RUBY-2023 Remove redundant apdex_s recording / assertions
Due to an artifact of how these tests used to be written, apdex_s,
metrics would be recorded twice for a transaction. This is because
record_apdex was being called directly in the test, rather than
allowing it to be called by Transaction#commit as it would under
realistic circumstances. The in_transaction block was the cause
the additional apdex_s metric. The tests had the following comment
explaning the situation:

  apdex_s is 2 because the transaction itself records apdex

Now that these test cases have been expanded out and apdex is recorded
as a result of Transaction#commit!, we can remove the comment and the
transactions that record the additional apdex_s metrics.