RUBY-1724 Remove frame_stack from Transaction
The frame_stack was used to match calls to the Transaction.start with
calls to Transaction.stop so that there could be multiple calls to start
so long as there are at least as many calls to stop. When the stack is empty
it's signal that the transaction is really finished. It was also used to
conditionally add the nested prefix to segments. We have replaced this machinery
by having Tracer.start_transaction or Tracer.start_transaction_or_add_segment return
an instance of a finishable, on which the finish method can be called. For naming
we use the nesting_max_depth method on Transaction.