RUBY-1724 Allow start_transaction_or_segment to take a name or partial_name and category
This commit allows either a name or partial_name and category to be
passed to start_transaction_or_segment. From the agent side, being
able to pass in a name, constructed as you want it, is useful and
reasonable. To make the API more approachable for third parties, being
able to pass in a partial name and a category will spare users from
having to know the details of our transaction naming conventions.