RUBY-1724 Replace double-splat with a single keyword argument
With the previous definition of `start_transaction_or_segment` taking
a double-splat `**options` argument, we were seeing a 12% increase in
allocations and 3% increase in time spent on the basic middleware
benchmark (compared to the old `Transaction.start/stop` API).

This commit replaces the argument with an `options:` keyword argument
defaulting to an empty hash, resulting in no performance loss or
allocation bloat for the same benchmark.