RUBY-2092 Use Tracer.current_transaction convenience method
There are a number of situations where code could be improved using
this method, but these were easy to find and change. Most of the changes
consist of replacing lines that fetch the state, then reference the
state for the current transaction, to skip the middleman and use
Tracer.current_transaction.