RUBY-2120 Notify in logs that the agent is disconnecting or restarting
We currently log exception info when the agent shuts down or restarts
because of the response it gets from the collector.  However, the log
entry does not actually say that a disconnect or restart is happening.

This commit updates the existing log messages to include the word
"disconnecting" or "restarting," respectively.