RUBY-2143 Cleanup test_helper; fallback to NO_RAILS when config/environment absent
This commit cleans up and removes code that was in a rescue block for
a LoadError that essentially had the contents of what is in
config/environment.rb for all Rails versions we have to today. Spelunking
through the git history shows that it was used for older, now unsupported
Rails environments, 2.3 for example. The rescue was repurposed for
falling back to a NO_RAILS mode when there is not a file at
config/environment.rb.