RUBY-2145 Fix TransactionTimeAggregator for JRuby
ObjectSpace is disabled by default on JRuby and was causing the TTA to
fail. Unit tests had been in allowed failures after they started failing
intermittently for indeterminate reasons. A recent upgrade on CI that
brought in a new JDK seems to have resovled the inexplainable failures
with this new, very explainable and fixable one.