RUBY-2145 Run units on JRuby in CI
Our JRuby tests started failing inexplicably, and for unreproducible
reasons on Travis after an AMI update. I recently updated the JDK,
JRuby version, and fixed a legitimate test failure on JRuby, and now
everything passes, and so far, they pass repeatedly. This commit
removes units from allowed failures, and adds the rails 4.0 and 4.1
environments to the exclude list. This is because JRuby 9.2.x.x
is Ruby 2.5 compatible, and Rails 4.0 and 4.1 ar not.