RUBY-2208 Allow enabled to be derived from multiple keys or fn
Our aggregators need to be able to derive their enabled state off of multiple
config keys, and watch those keys for changes, and recompute their status
when they do. There is also an option to provide an enabled fn for situtations
where the default &&ing of keys is not sufficient.