RUBY-2208 Make error event collection abide by error_collector.enabled
Previously it was possible, and in some ways "a feature", that error events could
be enabled and controlled separately from error traces. In order to make this work
as expected, error events should abide by the error_collector.enabled configuration.
It's possible to collect only error traces, but not events, but not the other way
around.