CI: Ubuntu 20.04
**Provide MySQL 5.5 from source**
For older (< v2.4) Rubies and the Rubies that are still tested against
Rails environments via `mysql2` v0.3.x (Ruby < 2.6), provide an instance
of MySQL 5.5 on the machine built from source and then cached for
subsequent runs. The newer MySQL versions require a higher version of
`mysql2` and the older (< 2.5) Rubies require OpenSSL v1.0, which MySQL
5.5 comes bundled with.

**Upgrade Ubuntu 18 to 20**
Upgrade from GitHub's 18.04 containers to 20.04

**Use MySQL as a service**
Ignore the MySQL server available (but not started) on the GitHub Ubuntu
20.04 image and use a separate service for MySQL instead. This brings
MySQL on a peer level with Postgres, Redis, etc.