Curb instrumentation method_with_tracing fix
To address an issue apparently dating back to when the Curb
instrumentation was first split into separate "prepend" and "chain"
approaches, be sure to pass a verb argument to `method_with_tracing`,
which requires it.

Thank you to @knarewski for bringing this issue to our attention,
for providing a means of reproducing an error, and for providing a fix.

resolves #1033