stdout/stderr log forwarding skipping tweaks
* In `NewRelic::Agent::Instrumentation::Logger`, move the registration
call to `#skip_instrumentation`
* Instead of clearing non standard stream loggers for instrumentation,
use a new dedicated @newrelic_registered instance variable to determine
whether a logger has been registered previously or not.
* When skipping standard stream loggers, simply ask the underlying
device if it's writing to a TTY instead of triple-equals checking if it
is exactly $stdout or $stderr.
* Don't bother to prune the registered loggers hash, which simply holds
references and is only used when new loggers are instantiated. Instead,
allow the hash to hold ALL loggers we've seen - skipped or cleared.
* Add a new test to make sure that a duped IO device based logger will
correctly be marked for skipping.