lower min compression threshold to 2048 bytes
lower the compression byte threshold:
`> 64 KiB`  ->  `>= 2 KiB`

previously the agent had a minimum byte size threshold of 64KiB. for any
payload equal to or below that size, compression would not be applied.
this relatively high threshold was originally established to prevent
segfaults from taking place in older versions of Ruby trying trying to
wrap C zlib/gzip libraries to compress small payloads.

with the currently supported Rubies (2.2 - 3.1), segfaults aren't a
problem for even the tiniest of payloads, so we are free to come up with
a value that balances cpu usage and bandwidth savings.