gRPC: define all individual method args
the gRPC client methods for sending data all accept an optional
`metadata` keyword argument. Because this argument is optional, the
traditional approach of prepending a method that accepts `*args` as
input is problematic given that the arguments aren't guaranteed to
contain the `metadata` keyword. We need to add additional key/val pairs
to the metadata hash, at all times - both when the caller supplied
metadata, and when it is missing and defaults to an empty hash. Given
our instrumentation's approach of calling `super` within a block, we
need to ensure that modified metadata always reaches `super` as input.
Switching away from the `*args` approach to explicitly defining all
input params allows us to accomplish this.

Many thanks to @tannalynn for her help in discussing this approach.

Resolves #1239