Segments: rework child timing range merging
For a given segment's children, each child's start and finish time pair
needs to be checked against every other child's pair of the same to
discover overlap and merge the timings when overlap is found.

Searching an array of ranges for an entry that overlaps with a given
range can be quite performance heavy, even when directly leaning on
`Range#cover?`. To avoid the performance penalty as much as possible, do
not incur it every time a new child's timings pair is seen. Instead,
keep the children timing pairs as 2 item arrays for as long as possible
without looking for overlap or merging any values. Then, only when it
comes time to record the timing, incur the hit a single time (and
memoize the result) to produce an array of ranges whose overlapping
entries have all been merged.