enable CLM by default
following the successes reported by a good number of opting-in adopters,
the decision has been made to enable code level metrics reporting in the
Ruby agent by default. users may still come up with feature requests for
additional CLM functionality or instrumented library coverage, but the
core functionality in place currently is thought to be stable and ready
for enabling by default.