Create TestTimeReporter
The TestTimeReporter outputs the ten slowest tests from the
previous run.
When running in the unit tests, it will print at the end of
the test run.
When running in multiverse, a file is created within
lib/minitest that holds the test output from each suite
and organizes the results at the end of all suites before
printing out the ten slowest results.