CI fixes for Ubuntu 22 compatibility
• Use a GHA env var for JAVA_HOME when setting alternates, but then
switch to a plain text string when running tests. This is an odd one
that was working previously with the use of environment variables
everywhere previously. For now we are avoiding having to bring in the
additional dependency and spend the additional processing time of
`setup-java` given that our desired JVM is on the machine already.
• Further rework the Bundler setup scripts we use for CI. Ruby 2.2 is
the only Ruby that we support that needs `update_rubygems` and the
others can all make use of `gem update --system`.
• Don't test Rails 3.2 on Rubies above v2.4 (remove the tests from the
v2.5 track).
• Don't test `redis` gem v5 yet until we determine why it throws CI
errors
• RuboCop syntax updates

3 2 1... let's Jammy Jellyfish

"Angels banished from heaven have no choice but to become devils."