Gabby: report only stable versions for 'previous'
When Gabby notifies us about a new gem version that has recently become
available, it would be preferable to:

- mention rc1/beta/preview releases for the newly released version to
  give the team a heads up about development work being done on the gem
- only report stable versions for the previously released version.

This way, if a gem goes `v1.5 -> v2.0.preview -> v2.0.rc1 -> v2.0.rc2`
then the notifications will always tell us about the previous stable
version, not the previous beta version.

Currently, the notifications look like this:

- v1.5 to v2.0.preview
- v2.0.preview to v2.0.rc1
- v2.0.rc1 to v2.0.rc2

With the included changes, the notifications would instead read:

- v1.5 to v2.0.preview
- v1.5 to v2.0.rc1
- v1.5 to v2.0.rc2

This way we are notified of beta releases and don't have to manually
search for the previous stable version.