Create string to insert global into template
The newrelic.yml file template uses ERB tags to interpolate a global
variable as the value to the instrumentation method being tested
(ex. chain or prepend).
ERB tags are also used by .tt files to interpolate code. This caused the
earlier implementation to attempt to find the $instrumentation_method
variable and add its value to the .yml file, causing a blank value.
Assigning the snippet we want injected into the yml file as an instance
variable and interpolating that value resolved the problem.