Move minitest_time_report reset to runner
The minitest_time_report is created by the TestTimeReporter class,
however, because minitest is loaded by multiverse suites,
trying to reset the report by calling a method in TestTimeReporter
loaded multiple minitest versions.
By resetting the file in the Multiverse::Runner#run method, we avoid
the need for the class to be loaded until the suites load minitest.