Update Redis instrumentation to support v5
Co-authored-by: hramadan <hramadan@newrelic.com>

Redis gem 5.x comprises a major structural refactor for the library.

* Calls to methods like get and set are no longer routed through `call`,
but through a method named `call_v`
* The methods we instrument have been moved into a new library,
redis-client, a dependency of the redis gem.
* Prepending onto the new library's class (::RedisClient) for the
`call_v` and connect methods maintains consistent reporting behavior
with earlier versions of Redis.
* The location, behavior, and arguments of pipelined and multi
methods changed.
* In previous versions, pipeline and multi calls were routed through a
method named `call_pipeline`. There is a similar method in Redis 5.x
named `call_pipelined`. However, this method is not defined in
::RedisClient, preventing our prepend instrumentation from working.
::RedisClient defines `multi` and `pipelined` methods, but they do not
have access to the NoSQL statement as an argument or other instance
variable/method on the caller.
* For `call_pipelined`, we now leverage Redis 5.x's new instrumentation
API, `::RedisClient.register(module_name)`.
* This process introduces instrumentation too late to capture errors at
the segment level. Despite this shortcoming, this strategy is used for
`pipelined` and `multi` calls because it has access to the NoSQL
statement