Adopt the Redis 5.0 Middleware instrumentation API
redis-rb 5.0 includes the redis-client gem. The client gem took over
responsibility for a lot of the code we instrument and created an
Instrumentation API:

RedisClient.register(Middleware)

Where middleware is the module that includes the instrumentation.
By adopting this method, we will not use prepend or chain strategies
when the installed version of redis-rb is above 5.0.0

The only shortcoming we've found so far is that we cannot tell the
difference between a multi request and a pipeline request.
This might still be possible, or we may need to raise
an issue/contribute to redis-client.