Adopt hybrid approach for Redis instrumentation
Redis gem 5.x comprises a major structural refactor for the library.
The call method has become call_v. The methods we instrument have
been moved into a new library, redis-client, a dependency of the redis.
Prepending onto the new library's class for the call_v and
connect methods maintain
consistent reporting behavior with earlier versions of Redis.
The location, behavior, and arguments of pipelined and multi
methods changed.
Redis 5.x introduces a new option for adding instrumentation,
::RedisClient.register(module_name). This process introduces
instrumentation too late to capture errors at the segment level.
Despite this shortcoming, this strategy is used for pipelined and
multi calls because it has access to the nosql statement,
whereas our common pattern of chaining/prepending the pipeline
or multi method call does not.
In earlier versions of Redis, pipeline and multi methods were
routed through call_pipeline, which was used for instrumentation.