Define all agent agent modules in agent class
NewRelic::Agent::Connect already existed when refactoring the connect
module out of agent agent. Because of this I used
NewRelic::Agent::Agent::Connect when creating the connect file. For
consistency, I've updated all other modules that are being treated the
same way as Connect to be defined inside of the agent class.