CI: test with Ruby 3.2.0-RC1
... So this is 3.2
And what has been done?
Another year over
And a new one just begun
And so this is 3.2
I hope you have fun
The devs and the hobbyists
The old timers and noobs

... A very Merry Ruby Release
And a Happy New Year
Let's hope it's a good one
Without any tears

... And so this is 3.2 (ooh)
For WASI based Wasm (ooh)
For large and small projects (ooh)
This release should be awesome (ooh)
And so happy 3.2 (a new release is here)
For CRuby and JRuby (if you've got Java)
For Truffle and Rubinius (use what you like)
Let's all get moving (now)

... A very Merry Ruby Release
And a Happy New Year
Let's hope it's a good one
Without any tears

... And so this is 3.2 (ooh)
For improved regex with timeouts (ooh)
For error highlights and suggestions (ooh)
And better performance no doubt (ooh)
And so happy 3.2 (a new release is here)
For CRuby and JRuby (if you've got Java)
For Truffle and Rubinius (use what you like)
Let's all get moving (now)

... A very Merry Ruby Release
And a Happy New Year
Let's hope it's a good one
Without any tears

... And so this is 3.2
And what has been done?
Another year over
And a new one just begun
And so this is 3.2
I hope you have fun
The devs and the hobbyists
The old timers and noobs

... A very Merry Ruby Release
And a Happy New Year (Ooh, oh)
Let's hope it's a good one. It's a good, it's a good one.
Without any tears

... And so this is 3.2 (upgrade when you like)
And what has been done? (if you want it)
Another year over (a new release is here)
And a new one just begun (now)