Enable defer_rails_initialization config by environment variable only
Configs are loaded during initialization. This fix was placed behind a
feature flag referencing the config value, but that value will always
be falsey at initialization, so the bugfix was unreachable.

Environment variables are accessible before initialization.
The environment variable NEW_RELIC_DEFER_RAILS_INITIALIZATION may be
set to true to enable this option.