removed check current txn check from thread block
Turns out we do need to copy the current transaction over every time.
Concurrent ruby can create a thread prior to the transaction beginning
or during a different transaction, but continue to use the same thread.
This ensures the agent is always using the most accurate source
for the current transaction