Add Active Support instrumentation
Active Support has notifications built in to instrument
caching-related activities, including:
* cache_read
* cache_generate
* cache_fetch_hit
* cache_write
* cache_delete
* cache_exist?

The payload always includes the key and the name of the
store class. cache_read also includes whether the read
is a hit, and :super_operation when a read is used with

https://guides.rubyonrails.org/active_support_instrumentation.html#active-support

When these operations are executed, a new segment will
be started that includes the payload within the params
attribute.