NoticedError: update attr cache for error group
`NewRelic::Agent::ErrorCollector#create_noticed_error`, which is the
sole consumer of `NoticedError#error_group=`, ends up calling
`NoticedError#processed_attributes` fully 3 times in order to build the
hash that is passed to the customer error group callback.

As a result, the agent attributes for the noticed error end up being
memoized. Because of the memoization, the agent attributes hash is not
rebuilt later on when it comes time to submit those attributes to the
collector.

Updated `NoticedError#error_group=` to effectively bust the memoization
cache and ensure that the error group name makes it into the agent
attributes hash.