test:multiverse:gem_manifest - go faster
The `test:multiverse:gem_manifest` rake task has been sped up so that it
instead of taking ~2.5 minutes, it'll only take ~3.5 seconds.

This speedup is achieved by not concerning ourselves with twiddle wakka
based versions that we are already aware of (considering them to be
specific enough already) and simply taking a single trip to RubyGems.org
via Bundler in order to fetch all of the "latest" version numbers for
each core tech gem.