CI: ignore Rails environment switches for db tasks
When switching between the default_env and test Rails environments, the
Rails database tasks will raise
`ActiveRecord::EnvironmentMismatchError`.

Rails can either be explicitly told that the environment being switched
to is expected via a task that sets the target environment, or the check
itself can be disabled via an environment variable.

For now we are opting for the latter with
`ENV['DISABLE_DATABASE_ENVIRONMENT_CHECK']`.