CI: remove test:multiverse:self
Remove the `test:multiverse:self` Rake task and its related test-unit
test code and runner script.

This task has been broken for quite some time, owing to the fact that it
requires test-unit which is no longer available as part of the default
installation for modern Rubies.

We could declare test-unit as a dev dependency to restore the tests, but
of the 3 tests 2 of them will fail because of the way we have reworked
Multiverse testing with things like SimpleCov and the reporting of the
slowest tests. The Rake task's tests expect the Multiverse process
itself to exit with a non-zero code in the event of a test failure and
it no longer does so by design.

resolves #1163