Include `test/agent_helper.rb` in build
The test/agent_helper.rb file was overlooked in our switch to the new
file inclusion process introduced in PR#2089
https://github.com/newrelic/newrelic-ruby-agent/pull/2089

This file is required for the public API
NewRelic::Agent.require_test_helper.

Furthermore, build.rb was still included. This file used to be
generated with Jenkins during our release process and has not been
created with a release since version 6.13.0.