Update elasticsearch port_path_or_id value
Previously, the port_path_or_id value for Elasticsearch segments
was set to the path given to the request_with_tracing method. This
caused a metrics grouping issue for a customer because a new
instance metric (ex. "Datastore/instance/Elasticsearch/localhost/*") was created for
every Elasticsearch document ID.

The hash used to set the host value has a port key. This PR updates the
value of port_path_or_id to use the port from the nr_hosts hash.