Autostart denylist: permit 'test' and Rake commands
The agent already has conventions in place for using a separate 'test'
environment specific configuration for test running, and for denying
specific Rake tasks by name, so do not redundantly disable them by
default with the constants denylist. This way, users who wish to use the
agent with their tests and Rake tasks can continue to do so.