Skip Rails 7.1 ActiveSupportLogger instrumentation
The private method ActiveSupport::Logger.broadcast has been removed
in Rails 7.1. This throws an error in our tests. It also prevents the
@skip_instrumenting instance variable from being set on broadcasted
loggers. This may cause users of Rails 7.1 to send duplicate
log events to New Relic.

This change stops ActiveSupport::Logger instrumentation from being
installed on Rails 7.1 or above. This is intended to be a temporary
measure until a solution can be found to stop sending duplicate logs
in Rails 7.1.