Ethon: disable when Typhoues is present
After some extensive testing (thanks, @kaylareopelle) we've decided to
disable Ethon instrumentation in the presence of Typhoeus. We already
create `Typhoeus/GET` segments for each instance of `Typhoeus::Request`
and seeing an additional `Ethon::Easy/GET` segment really just seems
to duplicate the number of HTTP requests the app performed. Given that
all of the information from an `Ethon::Easy/GET` segment is already
contained in the `Typhoeus/GET` segment, it's better to prevent the
duplication.