stop testing 8T with Ruby < 2.7
Infinite Tracing depends on the `grpc` gem, which requires Ruby 2.5+.

Frequently, the gRPC Ruby gem maintainers inadvertently break
compatibility with Ruby 2.5 and 2.6 by publishing a new version of
`google-protobuf` (which `grpc` depends on) that itself requires Ruby
2.7+.

We do not wish to pin our `grpc` dependency to an exact version that is
known to be compatible with Ruby 2.5 and Ruby 2.6 so that we leave users
free to use whatever 1.x version of `grpc` that they wish, especially if
they are themselves using `grpc` in their application.

Whether the gRPC maintainers' frequent breaking of Ruby 2.5 and 2.6
compatibility is accidental or deliberate, supporting these Rubies with
our CI system has simply caused too many issues.

We are not actively introducing any new Ruby code to Infinite Tracing
that could break Ruby 2.5 and 2.6 compatibility, so we will switch to
manual testing of these Rubies as appropriate moving forward.