don't log the license key even in debug/audit
By default the license key isn't logged given a) the default non-debug
log level, and b) the default of not enabling the audit log.

While accuracy in those 2 non-default contexts can be useful in
general, the inclusion of the license key does not provide much value
and can serve to make log data more sensitive than it needs to be.

Filter out the license key from all contexts.