Create :'elasticsearch.perform_health_check' config
To access the cluster name to assign to the
database_name attribute on the Elasticsearch datastore segment,
we currently make a request to the '_cluster/health' endpoint. This can
cause CPU and network spikes when large clusters start.
To avoid this performance issue, the
`:'elasticsearch.perform_health_check` config value can be set to false.
This will set the database_name attribute on the Elasticsearch datastore
segments to nil.