Remove error suppression in `render_in`
We should be able to se error pages, specially in when Rails.env is
development. The piece of code that was removed here
prevented from happening. Instead, it would have the `render_in` method
return `nil` when an error happened, resulting in either a blank page,
or an incomplete page.

I believe we also want to raise errors in production. We should not make
a decision here, whether the error should be suppressed. Otherwise we
are making the assumption "incomplete pages are better than error
pages", which may be the correct assumption under some circumstances,
but not under other circumstances.

Fixes https://github.com/ViewComponent/view_component/issues/1981