Create `token_count` attribute (#2489)
The token_count attribute is available on LlmChatCompletionMessage and LlmEmbedding events.

When a customer provides a callback using NewRelic::Agent.set_llm_token_count_callback,
that callback will be used to calculate the token count. Otherwise, the backend will attempt
to calculate the token count using the content and model. This callback is most likely to be
invoked when ai_monitoring.record_content.enabled is set to false.

token_count will not be assigned by the callback if the return value is not an integer greater
than zero.