simplify tests for AI header setting / not setting
use a variable in combination with a proc to confirm the stubbed method
was indeed invoked as expected or not invoked as expected.

this gives us an assertion in the non-invoked path, which is nice.

i believe this fixes (switching from `-> {}` to `proc { |_h| }` an issue
to where if the `raise` actually was attempted in the non-invoked path,
it would fail.

removed unused variables / mocks