Configuration: support allowlists
Configuration hashes defined by `default_source.rb` now support an
additional `:allowlist` key set equal to an array of valid values.

If an invalid value is specified for a parameter that has an allowlist,
a warning will be logged that contains the allowlist and the default
value for the parameter will be used.