CI: remove old Rack v1.x tests
We have some tests that were designed to work with Rack v1.x as long as
Rack was at or above v1.4. We checked to see if the minor version was at
or above 4 without checking the major version. These tests haven't
worked properly since Rack v2 as a result. Given that we no longer test
against Rack v1.x, the tests have been removed.