CI: Grape/Rack compat fixes
We need to give Rack <= v3.0 to Grape v1.x.

We may have a Grape version string in play, or it might be nil. When it
is nil, we can count on Grape 2 being used for Ruby 3+ and Grape 1 being
used for Ruby 2. So key off of both the Grape version and Ruby version
when deciding to force Rack v3.0.