unit tests: scope to files, favor the shortest
I invoked `bert agent`, with a desire to run the tests located at
`test/new_relic/agent_test.rb`. Instead the tests located at
`test/new_relic/agent/rpm_agent_test.rb` were ran instead. This is
because we were previously simply grabbing the very first `find`
command.

Now, all results obtained by the `find` command will be considered, and
the one with the shortest file path will be chosen.

In future we may want to detect the presence of `fzf` and prompt the
human tester when multiple results are found, but for now this seems to
yield pretty desirable behavior while still permitting the human to add
more characters to the input string if they wish to disambiguate between
multiple matches themselves.

Also, the `find` command was updated to only focus on files and not
directories.