aws-sdk-lambda fix
commit 0a1527d8012cfb34c3950d34180d0974b51ab9ed sought to improve error
handling, but incorrectly misused an `ensure` block to attempt to always
return the raw client response.

this faulty logic has been fixed. internal agent response handling
errors are logged as desired, not re-raised to the user's app, and the
user's app is guaranteed to always get the raw original response.