aws-sdk-lambda: don't use external segments
the use of external request segments can cause confusion for Lambda
entity relationships. should we wish to expose the underlying net call
to the AWS back-end API in future, we can do so by simply removing the
`disable_all_tracing` wrapper and allow the underlying Net::HTTP call to
be instrumented.