Add support for four sampler options
default, adaptive, always_on, always_off are now available for use when
a trace is root, remote_parent_sampled or remote_parent_not_sampled.

default is first in the case list because it is the most common. I
predict adaptive will be far less common than always_on or always_off,
so I put it last for a slight performance benefit.